refactor(share): align the split control with the share branch
Pull the reworked control from `claude/share-split-copy-button`: the
geometry moves into a standalone `SplitShareButton`, the copy glyph swap
into `CopyStateIcon`, and `ShareActions` composes them.

The dropdown changes with it. The "Share" heading goes — the control it
drops from is already labelled, so it only repeated it. The list is a
4-column grid rather than `flex-wrap`, so the columns divide the content
box exactly and the padding reads equal on all four sides, and the
padding needs `!p-4` to beat the dropdown's own `p-1.5` at equal
specificity. The chevron loses `pressed`: it renders `aria-pressed`,
which paints a held-down fill, and Radix already sets `aria-expanded`.

The two local deltas ride along unchanged — `shortenUrl` for the
referral link, and each half stopping its own click so the surrounding
field cannot steal focus and dismiss the menu.
